PoC pool of babels so transformation is less bottlenecked on them
this speeds up tc39 tests, but all other tests suffer.

It also has the nice benefit of dropping the babels once they are no
longer needed if you just run k6.

But a different approach is probably going to be needed as the js package
(for example) with -race takes upwards of 3x as much time.